requirements/base.txt
beanie==1.30.0
fastapi==0.115.13
langchain==0.3.26
langchain-community==0.3.26
langchain-mongodb==0.6.2
langchain-openai==0.3.24
langgraph==0.4.8
langgraph-checkpoint-mongodb==0.1.4
motor==3.7.1
pydantic-settings==2.9.1
pymongo==4.12.1
python-dotenv==1.1.0
python-multipart==0.0.20
uvicorn==0.34.3
PyJWT==2.10.1
passlib[bcrypt]==1.7.4
